Changed text input collection and dead key handling.
Use a newly discovered method to avoid changing the global dead key state.

If the dead key state must be changed, restore it ASAP.

The conditions for observing issues which this fixes are complicated, but one
example is shown below.  The end key is suppressed but is still buffered as a
dead key as a side effect of text translation by the hook:

    ; Confirmed on v2.0.17 with US-International layout.
    ih := InputHook('', '{vkDE}')  ; Press this key: '
    ih.Start
    ih.Wait
    InputBox ih.Input  ; Type a vowel in the InputBox.

Dead keys with the S and I InputHook options no longer affect input collection.

InputHook M option now permits Numpad digits to be collected while Alt is held,
and no longer contributes to Alt+nnnn input if the digit keys are suppressed.